Update to gulp 4
The tasks are now private functions that return streams which are composed
with gulp.series to form larger tasks.

Removed the use of the deprecated gulp-util and replaced it with direct
use of minimist and through2. The gulp-cssnano package has also been
deprecated but updating it is a bigger task and is better done in a
separate commit.

Added package-lock.json to git. This makes sure the correct versions of
dependencies are installed.