Allow package patch upgrades
A number of python packages use deprecated syntax internally which causes
warnings when tests are run.

Most packages should be allowed to do a patch upgrade to make sure we use
the version with the latest fixes but for now, only enable upgrades for
the ones which use deprecated syntax.