Process extra transaction information
If a transaction has a message attached to it, it is now included in the
NdaTransaction class. This will help identifty if a payment is made
without a reference but with a message.

The collection of the extra information for European bank transfers could
potentially be used to fix the problem where the reference numbers are
modified by the bank when payments are made from a foreign
(i.e. non-Finnish) bank account. The problem is that we currently have no
data to test this with so it has to be handled separately.