Add support for alternative CSV format
Adds rudimentary CSV parsing of payment records in the form of the
parseCSV function. Also added unit tests with the currently known format
of the CSV schema. Changes to the format are somewhat easy to do since
it uses DictReader.

Added a new button to the transactions admin page for uploading the new
CSV files. The old button for the NDA forms is left there for backwards
compatibility.

Fixed the parser.py unit test to use the python from the environment
instead of a hard-coded path. This makes it work with venv.